requirements.txt
absl-py==1.3.0
asttokens==2.2.1
backcall==0.2.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cryptography==38.0.1
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
executing==1.2.0
flit_core==3.6.0
fonttools==4.38.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna==3.4
importlib-metadata==5.1.0
ipython==8.7.0
jedi==0.18.2
joblib==1.2.0
kiwisolver==1.4.4
lap==0.4.0
Markdown==3.4.1
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
motmetrics==1.4.0
natsort==8.2.0
oauthlib==3.2.2
packaging==22.0
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
pip==22.3.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
rsa==4.9
scikit-learn==1.2.0
seaborn==0.12.1
setuptools==65.5.0
six==1.16.0
sklearn==0.0.post1
stack-data==0.6.2
tensorboard==2.11.0
termcolor==2.1.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tqdm==4.64.1
traitlets==5.7.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
xmltodict==0.13.0
yacs==0.1.6
zipp==3.11.0

# Yolo v7 requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation